requirements.txt
llama-index-core==0.14.10
llama-index-readers-file==0.5.5
llama-index-embeddings-huggingface==0.6.1
llama-index-llms-groq==0.4.1
sentence-transformers==5.2.0
python-dotenv==1.2.1
flask==3.1.2
gunicorn==23.0.0